Wonbin service and dao are merged to Havruta
