+++ src/main/java/com/example/havruta/data/dao/HavrutaDao.java
@@ -37,4 +37,16 @@ public interface HavrutaDao {
     List<CategoryProblemEntity> saveNewCategoryProblem(ProblemEntity problemEntity, List<CategoryEntity> categoryEntityList);
 
     void deleteProblem(Integer problemId);
+
+    public List<CategoryEntity> findAllCategory();
+    public Optional<UserEntity> findUserById(Integer userId);
+    public Optional<MemberEntity> findMemberById(Integer userId, Integer groupId);
+    public List<CategoryClosureEntity> findClosuresByRootId(Integer rootId);
+    public Optional<CategoryEntity> findCategoryById(Integer categoryId);
+    public List<MemberEntity> findMembersByGroupId(Integer groupId);
+    public Boolean designateAdmin(Integer newAdminId, Integer groupId);
+    public Boolean dropMember(Integer userId, Integer groupId);
+    public Boolean confirm(Integer userId, Integer groupId);
+    public Boolean deleteGroup(Integer groupId);
+    public Boolean updateGroup(String newGroupName, Integer groupId);
 }
