+++ src/main/java/com/example/havruta/service/HavrutaService.java
@@ -18,4 +18,15 @@ public interface HavrutaService {
     public ResponseDto updateProblem(String token, Integer problemId, ProblemDto problemDto);
 
     public ResponseDto deleteProblem(String token, Integer problemId);
+
+    public SpecificGroupResponseDto specificGroupPage(String token, Integer groupId);
+    public AdminResponseDto admin(String token, Integer groupId);
+    public AdminMembersResponseDto adminMembers(String token, Integer groupId);
+    public ResponseDto designateAdmin(String token, Integer newAdminId, Integer groupId);
+    public ResponseDto dropMember(String token, Integer userId, Integer groupId);
+    public ResponseDto confirm(String token, Integer userId, Integer groupId);
+
+    public ResponseDto deleteGroup(String token, Integer groupId);
+
+    public ResponseDto updateGroup(String token, String newGroupName, Integer groupId);
 }
